pyproject.toml
[tool.poetry]
name = "unreachable-mapper"
version = "0.1.0"
description = "A geospatial analysis tool for computing and visualizing the most unreachable locations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
rasterio = "^1.3.9"
geopandas = "^0.14.3"
shapely = "^2.0.3"
pyproj = "^3.6.1"
osmnx = "^1.9.1"
numpy = "^1.26.4"
scipy = "^1.12.0"
matplotlib = "^3.8.3"
contextily = "^1.5.0"
folium = "^0.15.1"
pyyaml = "^6.0.1"
click = "^8.1.7"
tqdm = "^4.66.2"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.2.0"
flake8 = "^7.0.0"
jupyter = "^1.0.0"

[tool.poetry.scripts]
unreachable-mapper = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
